requirements.txt
numpy==2.0.2
pandas==2.2.2
piexif==1.1.3
pillow==11.0.0
python-dotenv==1.0.1
pyyaml==6.0.2
tqdm==4.67.0
